platform-test-python/requirements.txt
httpx>=0.28.1
pytest>=8.3.4
pytest-asyncio>=0.25.3
pyyaml>=6.0.2
ruff>=0.9.9

